# Miscelleanous performance improvements