--- a/ctapipe/image/tests/test_timing_parameters.py
+++ b/ctapipe/image/tests/test_timing_parameters.py
@@ -100,3 +100,42 @@ def test_ignore_negative():
     assert_allclose(timing.slope, grad / geom.pix_x.unit, rtol=1e-2)
     assert_allclose(timing.intercept, intercept, rtol=1e-2)
     assert_allclose(timing.deviation, deviation, rtol=1e-2)
+
+
+def test_outliers():
+    from ctapipe.image import timing_parameters
+
+    """
+    Simple test that gradient fitting gives expected answers for perfect
+    gradient
+    """
+    grad = 2.0
+    intercept = 1.0
+    deviation = 0.1
+
+    geom = CameraGeometry.from_name("LSTCam")
+    hillas = HillasParametersContainer(x=0 * u.m, y=0 * u.m, psi=0 * u.deg)
+
+    random = np.random.RandomState(1)
+    x = geom.pix_x.value
+    peak_time = intercept + grad * x
+    peak_time += random.normal(0, deviation, geom.n_pixels)
+    peak_time[0] = 4
+    peak_time[10] = 20
+    peak_time[15] = 15
+
+    timing = timing_parameters(
+        geom,
+        image=np.ones(geom.n_pixels),
+        peak_time=peak_time,
+        hillas_parameters=hillas,
+        cleaning_mask=np.ones(geom.n_pixels, dtype=bool),
+    )
+
+    # Test we get the values we put in back out again
+    assert_allclose(timing.slope, grad / geom.pix_x.unit, rtol=1e-2)
+    assert_allclose(timing.intercept, intercept, rtol=1e-2)
+
+    assert np.isclose(
+        timing.deviation, np.std(peak_time - (grad * x + intercept)), rtol=0.01
+    )