## Quantity vs. Angle

```
In [4]: %timeit Angle(5, u.rad)                                                                                                               
10.1 µs ± 63.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [5]: %timeit u.Quantity(5, u.rad)                                                                                                          
3.46 µs ± 28.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

## Reimplement number_of_islands

From
```
318 µs ± 11.7 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
To
```
13.8 µs ± 15.4 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

## Timing parameters using sigma clipping instead of siegelslopes:

1855 pixels:

From
```
153 ms ± 382 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
To
```
311 µs ± 88.9 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

100 Pixels:
From 
```
4.35 ms ± 5.39 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
To
```
229 µs ± 102 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


## Overall

On the LST Muon Mono run:

From
```
10000/10000 [02:29<00:00, 66.99ev/s]
```
To
```
10000/10000 [01:28<00:00, 113.16ev/s]
```