@MaxNoe 2020-05-22
Okay, I am no bumping my head against this since an hour. Sometimes numba just get's stuck on `set.add`, I suspect a numba bug, but I cannot reproduces this outside of the num islands code