@MaxNoe 2020-05-25
For my contrived test case, the sigma clipping and the siegelslopes are very similar:
```
import numpy as np
import matplotlib.pyplot as plt
from scipy.stats import siegelslopes
from ctapipe.image.timing import linear_regression, sigma_clipping_linreg


np.random.seed(0)

x = np.random.uniform(-0.2, 0.2, 50)
x.sort()
y = 10 * x + 5 + np.random.normal(0, 0.2, 50)

y[-3] = 10
y[-5] = 12
y[-15] = 7


xp = np.array([x.min(), x.max()])

plt.plot(x, y, 'o')
plt.plot(xp, 10 * xp + 5, '-', label='truth')


a, b, cov = linear_regression(x, y)
plt.plot(xp, a * xp + b, '-', label='linear regression')

a, b, cov = sigma_clipping_linreg(x, y)
plt.plot(xp, a * xp + b, '-', label='sigma clipping')

a, b = siegelslopes(y, x)
plt.plot(xp, a * xp + b, '-', label='siegel slopes')


plt.legend()
plt.savefig('robust.png', dpi=300)

```

![robust](https://user-images.githubusercontent.com/5488440/82821475-7734de00-9ea4-11ea-8a10-e504a59eed10.png)