@kosack 2020-05-25
numba does support `np.linalg.lstsq` (https://numba.pydata.org/numba-doc/dev/developer/autogen_numpy_listing.html?highlight=lstsq#numpy.linalg.lstsq),  so you might use that rather than your own implementation (to avoid inverse problems), or use SVD or something