@MaxNoe 2020-05-25
But numpy.linalg.lstsq does not support weights, does it?