@kosack 2020-05-25
https://stackoverflow.com/questions/27128688/how-to-use-least-squares-with-weight-matrix