@MaxNoe 2020-06-09
Closed as it was getting to large and mixed changes affecting the results with changes only speeding things up. I will open follow-op PRs for the single changes.